third_party/WebKit/LayoutTests/fast/borders/border-hittest-expected.txt


If this test passes, area outside is body element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.nodeName is "BODY"
PASS successfullyParsed is true

TEST COMPLETE
